Orbit 2 cleanup and additional tests
