+++ orbit-2-test/src/main/java/com/babylon/orbit2/OrbitGiven.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Babylon Partners Limited
+ * Copyright 2020 Babylon Partners Limited
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@
  *  limitations under the License.
  */
 
-apply<JacocoPlugin>()
+package com.babylon.orbit2
 
-tasks.withType<JacocoReport> {
-    reports {
-        html.setEnabled(true)
-    }
+class OrbitGiven<HOST : Host<STATE, SIDE_EFFECT>, STATE : Any, SIDE_EFFECT : Any>(
+    private val host: HOST,
+    private val initialState: STATE
+) {
+    fun whenever(invocation: HOST.() -> Unit) =
+        OrbitInvocation(host, initialState, invocation)
 }
-
-tasks.getByName("test").finalizedBy(tasks.getByName("jacocoTestReport"))
