+++ orbit-2-test/src/main/java/com/babylon/orbit2/TestContainer.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2020 Babylon Partners Limited
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package com.babylon.orbit2
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.runBlocking
+
+internal class TestContainer<STATE : Any, SIDE_EFFECT : Any>(
+    initialState: STATE,
+    private val isolateFlow: Boolean
+) : RealContainer<STATE, SIDE_EFFECT>(
+    initialState = initialState,
+    settings = Container.Settings(),
+    orbitDispatcher = Dispatchers.Unconfined,
+    backgroundDispatcher = Dispatchers.Unconfined
+) {
+    private var dispatched = false
+
+    override fun <EVENT : Any> orbit(
+        event: EVENT,
+        init: Builder<STATE, SIDE_EFFECT, EVENT>.() -> Builder<STATE, SIDE_EFFECT, *>
+    ) {
+        if (!isolateFlow || !dispatched) {
+            dispatched = true
+            runBlocking {
+                collectFlow(
+                    event,
+                    init
+                )
+            }
+        }
+    }
+}
