+++ orbit-2-test/src/main/java/com/babylon/orbit2/TestStreamObserver.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2020 Babylon Partners Limited
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package com.babylon.orbit2
+
+class TestStreamObserver<T>(stream: Stream<T>) {
+    private val _values = mutableListOf<T>()
+    private val closeable: Stream.Closeable
+    val values: List<T>
+        get() = _values
+
+    init {
+        closeable = stream.observe {
+            _values.add(it)
+        }
+    }
+
+    fun awaitCount(count: Int, timeout: Long = 5000L) {
+        val start = System.currentTimeMillis()
+        while (values.count() < count) {
+            if (System.currentTimeMillis() - start > timeout) {
+                break
+            }
+            Thread.sleep(10)
+        }
+    }
+
+    fun close(): Unit = closeable.close()
+}
